requirements.txt
opencv-python
pypdf2
pillow